requirement.txt
django==3.0.7
mysqlclient==2.0.3
djangorestframework==3.12.4
requests==2.25.1
django-cors-headers==3.7.0
